+++ pages/index.js
@@ -2,12 +2,14 @@ import { getSession } from "next-auth/react";
 
 import HomepageHero from "../components/HomepageHero";
 import AboutAlertit from "../components/AboutAlertit";
+import GetStarted from "../components/GetStarted";
 
 export default function HomePage() {
 	return (
 		<main className="bg-indigo-700">
 			<HomepageHero />
 			<AboutAlertit />
+			<GetStarted />
 		</main>
 	);
 }