+++ CHANGELOG.md
@@ -30,6 +30,7 @@ All notable changes to this project will be documented in this file.
 - Homepage Hero component
 - About Alertit component with dummy text
 - React icons package
+- GetStarted Component
 
 ### Changed
 - Add Sign In Form component